# Optimized pooling using OpenMP