I would like to suggest changes introducing parallelization of Pooling Layer (CPU) of caffe. This is another set of changes aiming to optimize CPU path using OpenMP. First is here:
https://github.com/BVLC/caffe/pull/3358

Changes are independent from above mentioned PR.

#### Modifications:
* Pooling Layers Forward and Backward parallelized using OpenMP 
* Optimized caffe_set using OpenMP
* Extended Cmake and Make files with code enabling OpenMP if supported by compiler
* Fixed CMake files to detect MKL

## Performance Results :
Presentated data was taken from caffe linked against MKL, as visible speedup is more significant that when using other BLAS libs (OpenBLAS, Atlas)
This is due to fact that in Atlas, OpenBLAS GEMM code takes more time than in MKL and optimization of pooling itself is not bringing that much speedup as 
when using MKL. It does help of course for all BLAS libs that  caffe building system supports.
Time measures are: average Forward-Backward as stated by *caffe time*. *speedup* is (master branch measure / openmp-pooling branch measure)

#### Intel(R) Xeon(R) CPU E5-2699 v3 @ 2.30GHz (MKL 11.3, GCC 4.8.3):
Branch | googlenet(speedup: 1.21(HT), 1.74(no HT)) | caffenet (speedup: 1.05(HT), 1.19(no HT)) | alexnet(speedup: 1.07(HT), 1.19(no HT))
----------|-----------------------|-------------------------------|-----------------------------
openmp-pooling (no HT)            | 2543ms                   |6809ms                   |7212ms 
openmp-pooling (HT)               | 3647ms                   |7745ms                   |8025ms 
master                            | 4438ms                   |8164ms                   |8644ms 

####  Intel(R) Core(TM) i7-4770R CPU @ 3.20GHz(MKL 11.3, GCC 4.9.2):
Branch | googlenet(speedup: 1.03(HT), 1.40(no HT)) | caffenet (speedup: 1.00(HT), 1.12(no HT)) | alexnet(speedup: 1.00(HT), 1.10(no HT))
----------|-----------------------|-------------------------------|-----------------------------
openmp-conv-relu (no HT) | 2863ms | 8516ms    |  9150ms
openmp-conv-relu (HT)    | 3876ms | 9512ms    |  10080ms
master                   | 4025ms | 9568ms    |  10127ms   

So there is significant speedup, that depends on how many CPU cores the platform does have. Tests were made using MKL(it is available free of charge now).

- For disabled openmp build performance remain comparable to the current master branch. 
- **It is best to have HT disabled in BIOS as using OMP_NUM_THREADS may not always prevent
OS kernel for running two OpenMP threads on the same phyisical core (eg. using HT).** 
- Without using OMP_NUM_THREADS and with HT enabled performance is still better than single threaded version

### Building:
* Install Intel MKL
* Modify Makefile.config so it has enabled BLAS:=mkl
* cmake <path to CmakeLists.txt> -DBLAS=mkl
* cmake <path to CmakeLists.txt> -DBLAS=mkl -DUSE_OPENMP=OFF # To have OpenMP disabled 
 

### Running:
##### Example of running:
###### Xeon E5 , two sockets, 18 cpu cores in each socket (**simulation of not using HT**)
*OMP_NUM_THREADS=36 ./build/tools/caffe time -iterations 20  --model=models/bvlc_googlenet/train_val.prototxt*  
###### Xeon E5 , two sockets, 18 cpu cores in each socket (**using HT**)
./build/tools/caffe time -iterations 20  --model=models/bvlc_googlenet/train_val.prototxt*  
###### i7-4770R (Haswell), one socket, 4 cpu cores in each socket (**simulation of not using HT**)
*OMP_NUM_THREADS=4 ./build/tools/caffe time -iterations 20  --model=models/bvlc_googlenet/train_val.prototxt*   
###### i7-4770R (Haswell), one socket, 4 cpu cores in each socket (**using HT**)
*./build/tools/caffe time -iterations 20  --model=models/bvlc_googlenet/train_val.prototxt*   


If you have any questions, please do not hesitate to ask.

Regards,